+++ table/block_based/block_based_table_reader.cc
@@ -1253,8 +1253,12 @@ Status BlockBasedTable::GetDataBlockFromCache(
   Statistics* statistics = rep_->ioptions.statistics.get();
   bool using_zstd = rep_->blocks_definitely_zstd_compressed;
   const FilterPolicy* filter_policy = rep_->filter_policy;
-  Cache::CreateCallback create_cb = GetCreateCallback<TBlocklike>(
-      read_amp_bytes_per_bit, statistics, using_zstd, filter_policy);
+  CacheCreateCallback<TBlocklike> callback(read_amp_bytes_per_bit, statistics,
+                                           using_zstd, filter_policy);
+  // avoid dynamic memory allocation by using the reference (std::ref) of the
+  // callback. Otherwise, binding a functor to std::function will allocate extra
+  // memory from heap.
+  Cache::CreateCallback create_cb(std::ref(callback));
 
   // Lookup uncompressed cache first
   if (block_cache != nullptr) {
@@ -1284,8 +1288,11 @@ Status BlockBasedTable::GetDataBlockFromCache(
   BlockContents contents;
   if (rep_->ioptions.lowest_used_cache_tier ==
       CacheTier::kNonVolatileBlockTier) {
-    Cache::CreateCallback create_cb_special = GetCreateCallback<BlockContents>(
+    CacheCreateCallback<BlockContents> special_callback(
         read_amp_bytes_per_bit, statistics, using_zstd, filter_policy);
+    // avoid dynamic memory allocation by using the reference (std::ref) of the
+    // callback. Make sure the callback is only used within this code block.
+    Cache::CreateCallback create_cb_special(std::ref(special_callback));
     block_cache_compressed_handle = block_cache_compressed->Lookup(
         cache_key,
         BlocklikeTraits<BlockContents>::GetCacheItemHelper(block_type),